Hah - as I feared, the Mac moaned about my newer int128_t code. I believe this is to do with whether uint64_t and unsigned long long are just two names for the same type or whether they are different but happen to behave the same. This can lead to ambiguous overloads for a function that can accept either int64_t or uint64_t as an argument but is provided with 0xffffffffffffffffU which clang on the Mac thinks might map almost equally well onto either. But putting an explicit static_cast<uint64_t> in there I can make that overload clearly the best. This is all RIGHT on the edge of my understanding of C++ and how to be certain code is unambiguous in all cases however a particular compiler has implemented the standards!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5391 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/int128_t.h
@@ -1877,7 +1877,9 @@ inline uint128_t uint128(std::uint64_t v)
 }
 
 inline uint128_t uint128(int64_t v)
-{   return INT128::PACK128(v<0 ? 0xffffffffffffffffU : 0U, v);
+{   return INT128::PACK128(
+        static_cast<std::uint64_t>(v<0 ? 0xffffffffffffffffU : 0U),
+        v);
 }
 
 #ifdef HAVE_INT128_T
